backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Pydantic & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
cryptography==42.0.2

# PDF Processing
pdfplumber==0.11.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.2.0

# AI/ML
openai==1.10.0
langchain==0.1.4
tiktoken==0.5.2

# Storage (S3/MinIO)
boto3==1.34.27
botocore==1.34.27

# Redis (Caching + Sessions)
redis==5.0.1
hiredis==2.3.2

# Background Workers
celery[redis]==5.3.6
flower==2.0.1

# Rate Limiting
slowapi==0.1.9

# CORS & Middleware
python-cors==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
user-agents==2.2.0

# Audit & Security
itsdangerous==2.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
bandit==1.7.6
safety==3.0.1

# Monitoring (Optional)
sentry-sdk[fastapi]==1.39.2
